requirements.txt
fastapi==0.128.0 
uvicorn[standard]==0.40.0 

python-jose==3.5.0
python-dotenv==1.2.1 

scikit-learn==1.4.2
joblib
pandas
numpy==1.26.4

redis== 7.1.0 

prometheus-fastapi-instrumentator==7.1.0 